Cleanup.

+++ src/Unmanaged/Constants.cs
@@ -28,7 +28,7 @@ static class Constants
 
     internal const int FILE_SHARE_DELETE = 0x00000004;
 
-    internal const nint INVALID_HANDLE_VALUE = -1;
+    internal const int INVALID_HANDLE_VALUE = -1;
 
     internal const int FileStandardInfo = 1;
